refactor: improve file structure and organization
