--- src/indices/type_algorithm.rs
@@ -1,72 +1,0 @@
-use super::*;
-
-/// Algorithm options used to index and search vectors.
-#[allow(missing_docs)]
-#[derive(Debug, PartialEq, Eq, Clone)]
-#[derive(Serialize, Deserialize)]
-pub enum IndexAlgorithm {
-    BruteForce, // -> IndexBruteForce
-}
-
-impl IndexAlgorithm {
-    /// Initializes a new index based on the algorithm and configuration.
-    pub(crate) fn initialize(
-        &self,
-        config: SourceConfig,
-        metric: DistanceMetric,
-    ) -> Box<dyn VectorIndex> {
-        let index = match self {
-            IndexAlgorithm::BruteForce => IndexBruteForce::new(config, metric),
-        };
-
-        Box::new(index)
-    }
-
-    pub(crate) fn load_index(
-        &self,
-        path: impl AsRef<Path>,
-    ) -> Result<Box<dyn VectorIndex>, Error> {
-        match self {
-            IndexAlgorithm::BruteForce => {
-                let index = Self::_load_index::<IndexBruteForce>(path)?;
-                Ok(Box::new(index))
-            }
-        }
-    }
-
-    /// Persists the index to a file based on the algorithm.
-    /// - `path`: Path to the file where the index will be stored.
-    /// - `index`: Index to persist as a trait object.
-    pub(crate) fn persist_index(
-        &self,
-        path: impl AsRef<Path>,
-        index: &dyn VectorIndex,
-    ) -> Result<(), Error> {
-        match self {
-            IndexAlgorithm::BruteForce => {
-                Self::_persist_index::<IndexBruteForce>(path, index)
-            }
-        }
-    }
-
-    fn _load_index<T: VectorIndex + IndexOps + 'static>(
-        path: impl AsRef<Path>,
-    ) -> Result<T, Error> {
-        let index = T::load(path)?;
-        Ok(index)
-    }
-
-    fn _persist_index<T: VectorIndex + IndexOps + 'static>(
-        path: impl AsRef<Path>,
-        index: &dyn VectorIndex,
-    ) -> Result<(), Error> {
-        let index = index.as_any().downcast_ref::<T>().ok_or_else(|| {
-            let code = ErrorCode::InternalError;
-            let message = "Failed to downcast index to concrete type.";
-            Error::new(code, message)
-        })?;
-
-        index.persist(path)?;
-        Ok(())
-    }
-}
