+++ src/lib.rs
@@ -4,6 +4,8 @@
 #![doc(html_favicon_url = "https://i.postimg.cc/W3T230zk/favicon.png")]
 #![doc(html_logo_url = "https://i.postimg.cc/Vv0HPVwB/logo.png")]
 
+pub(crate) mod utils;
+
 /// Primary module for vector database operations.
 pub mod db;
 /// Module for managing database indices and related types.
