--- src/types/conn.rs
@@ -1,2 +1,0 @@
-pub use sqlx::AnyConnection as SourceConnection;
-pub use sqlx::Connection;
