+++ src/types/mod.rs
@@ -1,9 +1,8 @@
+/// Vector indexing distance metrics.
+pub mod distance;
 /// Custom error types of OasysDB.
 pub mod err;
+/// Native index filtering implementation.
+pub mod filter;
 /// Vector record types for indices.
 pub mod record;
-
-/// SQL database connection types.
-pub(crate) mod conn;
-/// File operation utilities.
-pub(crate) mod file;
